Update README.md
I am thrilled to share my experience of developing a E-Commerce full stack app for a online transaction of national means . Alhumduillah, my client was highly satisfied with my attractive UI and vast knowledge of widgets and packages. My code was smooth and clean with captivating ui. This was 3 months long project, but the waiting worth it. I along with my team were engaged in this project. And I'm committed to satisfy my clients in future with same spirit. This app is about sending and receiving money. you can also pay for online expenses, bills, or lend money to your friends. This makes life and marketing easy. Online transaction apps are the back-bone for all E-commerce marketing